+++ instruments/sim.py
@@ -1,194 +1,114 @@
-import pyvisa
-import datetime
+import random
+import time
 
 class Instrument:
     def __init__ (self, logger, config:dict):
         '''Opens instrument'''
-        logger.debug('Initializing instrument...')
-        try:
-            logger.debug('Looking for instrument...')
-            self.rm = pyvisa.ResourceManager('@sim')
-            self.srcmtr = self.rm.open_resource('GPIB::8::INSTR', read_termination='\n')
-        except:
-            logger.critical('Instrument not found!')
-            quit()
+        logger.critical('THIS IS A SIMULATED INSTRUMENT, NO HARDWARE IS BEING USED! "[SIM]" WILL BE PRINTED BEFORE ALL LOGS')
+        logger.debug('[SIM] Initializing instrument...')
+        self.voltage = 0.0
+        self.current = 0.0
+        self.resistance = 0.0
+        self.timer = 0.0
+        self.output_state = False
+        logger.info('[SIM] Instrument successfully initialized!')
 
-#Basic control commands -------------------------------------------------------
+
+    # Basic control commands -------------------------------------------------------
     
     def close(self, logger):
         '''Closes instrument'''
-        pass
+        logger.debug('[SIM] Instrument closed')
 
-    def init(self,logger):
+    def init(self, logger):
         '''Initializes instrument'''
-        try:
-            self.inst.write('INIT')
-            logger.debug('Instrument initialized')
-        except:
-            logger.critical('Instrument could not be initialized! Check connection!')
-            quit()
+        logger.debug('[SIM] Instrument initialized')
 
     def reset(self, logger):
         '''Resets instrument'''
-        pass
+        logger.debug('[SIM] GPIB defaults reseted')
 
-    def write(self, logger, command:str):
+    def write(self, logger, command: str):
         '''Writes command to instrument'''
-        pass
+        logger.debug(f'[SIM] Instrument command: {command}')
 
-    def query(self, logger, command:str):
+    def query(self, logger, command: str):
         '''Queries instrument'''
-        try:
-            query = self.inst.query(command)
-            logger.debug(f'Instrument query: {query}')
-            return query
-        except:
-            print('Instrument could not be queried! Check connection!')
-            quit()
+        logger.debug(f'[SIM] Instrument query: {command}')
 
-    def set_output_state(self, logger, state:str):
+    def set_output_state(self, logger, state: str):
         '''Sets output state'''
-        pass
+        if state == 'ON': self.output_state = True
+        elif state == 'OFF': self.output_state = False
+        else: logger.critical('[SIM] Invalid output state!')
+        logger.debug(f'Instrument output set to {state}')
 
-    def set_output_value(self, logger, func:str, value:float):
+    def set_output_value(self, logger, func: str, value: float):
         '''Sets output value of voltage or current'''
-        try:
-            match func:
-                case 'Current': pass
-                case 'Voltage': pass
-            logger.debug(f'Instrument value set to {value}')
-        except:
-            logger.critical('Instrument value could not be set! Check connection!')
-            quit()
+        if func == 'Voltage': self.voltage = value
+        elif func == 'Current': self.current = value
+        else: logger.critical('[SIM] Invalid output function!')
 
-    def get_output_value(self, logger, func:str):
-        '''Measures output value of voltage, current or resistance'''
-        #try:
-        match func:
-            case 'Current': 
-                out = self.inst.query('MEAS:CURR?').split(',')[1]
-            case 'Voltage': 
-                out = self.inst.query('MEAS:VOLT?').split(',')[0]
-            case 'Resistance': 
-                out = self.inst.query('MEAS:RES?').split(',')[2]
-        #except:
-        #    logger.critical('Instrument value could not be measured! Check connection!')
-        #    quit()
+        logger.debug(f'Instrument {func} value set to {value}')
+
+    def get_output_value(self, logger, func: str):
+        '''Measures output value of voltage, current, or resistance'''
+        # Simulated response based on the function
+        if func == 'Voltage': return self.voltage
+        elif func == 'Current': return self.current
+        elif func == 'Resistance': return self.resistance
+        else: logger.critical('[SIM] Invalid output function!')
 
     def fetch(self, logger):
         '''Fetches instrument and returns list of values'''
-        try:
-            list_query = self.inst.query('FETC?')[:-1].split(',')
-            logger.debug(f'Instrument fetch: {list_query}')
-            listd_return = []
-            for val in list_query:
-                listd_return.append(float(val))
-            return listd_return
-        except:
-            print('Instrument could not be fetched! Check connection!')
-            quit()
+        # Simulated response
+        logger.debug('[SIM] Instrument fetched')
+        return [self.voltage, self.current, self.resistance ,self.timer]
 
     def read(self, logger):
-        '''Reads instrument and returns list of sample values'''
-        return [-0.1999855,-2.64061e-11,0.05002385,-1.99749e-11,9.91e+37,3.756836,'2023-04-15 16:40:47']
+        '''Reads instrument and returns list of values'''
+        # Simulated response
+        logger.debug('[SIM] Instrument read')
+        return [self.voltage, self.current, self.resistance ,self.timer]
 
-
-    def get_timer(self,logger):
+    def get_timer(self, logger):
         '''Returns timer value'''
-        return datetime.datetime.now().second
+        # Simulated response
+        logger.debug('[SIM] Instrument timer read')
+        return self.timer
 
     def reset_timer(self, logger):
         '''Resets timer'''
-        try:
-            pass
-            logger.debug('Instrument timer reseted')
-        except:
-            logger.critical('Instrument timer could not be reseted! Check connection!')
-            quit()
+        logger.debug('[SIM] Instrument timer reseted')
 
-#Measurement mode commands ------------------------------------------------
+    # Measurement mode commands ------------------------------------------------
 
-    def set_mode_fixed(self, logger , func:str):
+    def set_mode_fixed(self, logger, func: str):
         '''Sets measurement mode to fixed'''
-        try:
-            match func:
-                case 'Current': pass
-                case 'Voltage': pass
-            logger.debug('Instrument measurement mode set to fixed')
-        except:
-            logger.critical('Instrument measurement mode could not be set! Check connection!')
-            quit()
+        logger.debug('[SIM] Instrument measurement mode set to fixed')
 
-#Measurement function commands -------------------------------------------
+    # Measurement function commands -------------------------------------------
 
-    def set_src_func(self, logger, func:str):
+    def set_src_func(self, logger, func: str):
         '''Sets source function of voltage or current'''
-        try:
-            match func:
-                case 'Current': pass
-                case 'Voltage': pass
-            logger.debug(f'Instrument source function set to {func}')
-        except:
-            logger.critical('Instrument source function could not be set! Check connection!')
-            quit()
+        logger.debug(f'[SIM] Instrument source function set to {func}')
 
-    def set_func_range(self, logger, func:str, range:str=':AUTO ON'):
+    def set_func_range(self, logger, func: str, range: str = ':AUTO ON'):
         '''Sets range of voltage or current'''
-        try:
-            match func:
-                case 'Current': pass
-                case 'Voltage': pass
-                case 'Resistance': pass
-            logger.debug(f'Instrument range set to {range}')
-        except:
-            logger.critical('Instrument range could not be set! Check connection!')
-            quit()
+        logger.debug(f'[SIM] Instrument range set to {range}')
 
-    def set_func_step(self, logger, func:str, step:float):
+    def set_func_step(self, logger, func: str, step: float):
         '''Sets step'''
-        try:
-            match func:
-                case 'Current': pass
-                case 'Voltage': pass
-            logger.debug(f'Instrument step set to {step}')
-        except:
-            logger.critical('Instrument step could not be set! Check connection!')
-            quit()
+        logger.debug(f'[SIM] Instrument step set to {step}')
 
-    def set_sense_func(self, logger, func:str):
-        '''Sets sense function of voltage, current or resistance'''
-        try:
-            match func:
-                case 'Current': pass
-                case 'Voltage': pass
-                case 'Resistance': pass
-            logger.debug(f'Instrument sense function set to {func}')
-        except:
-            logger.critical('Instrument sense function could not be set! Check connection!')
-            quit()
+    def set_sense_func(self, logger, func: str):
+        '''Sets sense function of voltage, current, or resistance'''
+        logger.debug(f'[SIM] Instrument sense function set to {func}')
 
-    def set_func_ccplc(self, logger, func:str, value:float):
-    
+    def set_func_cplc(self, logger, func: str, value: float):
         '''Sets compliance of voltage or current'''
-        try:
-            match func:
-                case 'Current': pass
-                case 'Voltage': pass
-            logger.debug(f'Instrument compliance set to {value}')
-        except:
-            logger.critical('Instrument compliance could not be set! Check connection!')
-            quit()
+        logger.debug(f'[SIM] Instrument compliance set to {value}')
 
-    def set_func_nplc(self, logger, func:str, value:float):
+    def set_func_nplc(self, logger, func: str, value: float):
         '''Sets integration time for current or voltage'''
-
-        try:
-            match func:
-                case 'Voltage': pass
-                case 'Current': pass
-            logger.debug(f'Instrument integration time set to {value}')
-        except:
-            logger.critical('Instrument integration time could not be set! Check connection!')
-            quit()
-
-
+        logger.debug(f'[SIM] Instrument integration time set to {value}')
